pyproject.toml
[tool.poetry]
name = "youtube"
version = "0.1.0"
description = "YouTube project"
authors = ["Umesh <umesh@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
moviepy = "^1.0.3"
google-api-python-client = "^2.183.0"
google-auth-oauthlib = "^1.2.2"
google-auth-httplib2 = "^0.2.0"
Pillow = "^10.0.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"